Time: 0 ms (100%), Space: 36.4 MB (21.94%) - LeetHub

+++ 1899-count-items-matching-a-rule/1899-count-items-matching-a-rule.cpp
@@ -13,7 +13,7 @@ public:
 
     int countMatches(vector<vector<string>>& items, string ruleKey, string ruleValue)
     {
-        int answer = 0;
+        unsigned short answer = 0;
 
         for (auto &item : items)
             if (item[um[ruleKey]] == ruleValue)
